feat(inclusion_connect): renaming www app
