+++ lacommunaute/www/inclusion_connect_views/tests.py
@@ -12,7 +12,7 @@ from lacommunaute.inclusion_connect.models import InclusionConnectState
 from lacommunaute.users.factories import UserFactory
 from lacommunaute.users.models import User
 from lacommunaute.utils.testing import reload_module
-from lacommunaute.www.inclusion_connect.views import InclusionConnectSession
+from lacommunaute.www.inclusion_connect_views.views import InclusionConnectSession
 
 
 OIDC_USERINFO = {
